# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    WorkDossierAssetEntry:
      type: object
      description: "One inlined figure: what it is, the URL and markdown line that\
        \ render it, and the pages that name it."
      properties:
        id:
          type: string
        kind:
          type: string
        mimeType:
          type: string
        label:
          type: string
        url:
          type: string
        markdown:
          type: string
        pageIds:
          type: array
          items:
            type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
    WorkDossierAssetInline:
      type: object
      required:
      - sourceId
      - kind
      description: A figure of the entity's refinement to copy into its dossier.
      properties:
        sourceId:
          type: string
          pattern: \S
          description: The sketch's or the design's id.
        kind:
          type: string
          pattern: \S
          description: IMAGE (a sketch) or DESIGN
    WorkDossierAssetInlined:
      type: object
      description: "What was inlined, and the markdown line that renders it."
      properties:
        id:
          type: string
        kind:
          type: string
        mimeType:
          type: string
        label:
          type: string
        url:
          type: string
        markdown:
          type: string
    WorkDossierAssetList:
      type: object
      description: "The figures a dossier inlines, by label."
      properties:
        assets:
          type: array
          items:
            $ref: "#/components/schemas/WorkDossierAssetEntry"
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://dev-qits-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
...
  /projects/api/epics/{epicId}/dossier-assets/{assetId}/content:
    get:
      summary: "The bytes of one figure an epic's dossier inlines, sandboxed"
      operationId: getDossierAssetContent
      parameters:
      - name: assetId
        in: path
        required: true
        schema:
          type: string
      - name: epicId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Dossier Asset Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
...
  /projects/api/work/{qualifiedId}/dossier-assets:
    get:
      summary: List the figures a work entity's dossier inlines
      description: "Each asset's kind, mime type, label, content URL, markdown line\
        \ and the pages that name it, by label. No bytes: getWorkDossierAssetContent\
        \ serves those."
      operationId: listWorkDossierAssets
      parameters:
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The figures
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkDossierAssetList"
        "404":
          description: "No entity with this id, or one whose dossier inlines nothing"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Work Dossier Asset Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
    post:
      summary: Copy a figure of the refinement into an epic's dossier
      description: "Copies a sketch or a design of the epic's own refinement in, keeping\
        \ its id, and answers the markdown line to paste. Inlining the same figure\
        \ again answers the same line. The path names the epic by qualified id (<projectSlug>-<n>)\
        \ or UUID."
      operationId: inlineWorkDossierAsset
      parameters:
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkDossierAssetInline"
        required: true
      responses:
        "200":
          description: The inlined figure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkDossierAssetInlined"
        "403":
          description: An agent outside its own project
        "404":
          description: "No entity with this id, one whose dossier inlines nothing,\
            \ or no such figure in its refinement"
        "401":
          description: Not Authorized
        "400":
          description: Bad Request
      tags:
      - Work Dossier Asset Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
  /projects/api/work/{qualifiedId}/dossier-assets/{assetId}/content:
    get:
      summary: "The bytes of one figure a work entity's dossier inlines, sandboxed"
      description: "Served under the row's own mime type with Content-Security-Policy:\
        \ sandbox, X-Content-Type-Options: nosniff and private caching, on every response."
      operationId: getWorkDossierAssetContent
      parameters:
      - name: assetId
        in: path
        required: true
        schema:
          type: string
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The figure's bytes
          content:
            application/json:
              schema: {}
        "404":
          description: "No such entity, or no such figure in its dossier"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Work Dossier Asset Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
...
info:
  title: qits-projects API
  version: 2026.1006.121753